+++ src/GLideNUI/ConfigDlg.h
